libs/requirements.txt
jupyter
cassandra-driver
pandas
numpy